Bearbeiten Blatt 09 Aufgabe 9.3

+++ inf1/src/blatt08/BasisWandler.java
@@ -49,7 +49,7 @@ public class BasisWandler {
 		if (n <= 0L) {// Rekursionsbasis
 			ret = ""; // Führende Nullen werden nicht angegeben
 		} else {
-			ret = inBasisKwandeln(n / k, (int) k) + intNachZiffer((int) (n % k));
+			ret = inBasisKwandeln(n / k, k) + intNachZiffer((int) (n % k));
 		}
 		return ret;
 	}
